+++ packages/i18n/src/types.ts
@@ -7,6 +7,7 @@ export type Language = {
     | 'fr'
     | 'ja'
     | 'kr'
+    | 'nb'
     | 'nl'
     | 'ru'
     | 'sl'