+++ website/src/routes/api/(types)/NotLengthValidation/properties.ts
@@ -1,0 +1,46 @@
+import type { PropertyProps } from '~/components';
+
+export const properties: Record<string, PropertyProps> = {
+  TInput: {
+    modifier: 'extends',
+    type: {
+      type: 'union',
+      options: [
+        'string',
+        {
+          type: 'array',
+          item: 'any',
+        },
+      ],
+    },
+  },
+  TRequirement: {
+    modifier: 'extends',
+    type: 'number',
+  },
+  BaseValidation: {
+    type: {
+      type: 'custom',
+      name: 'BaseValidation',
+      href: '../BaseValidation/',
+      generics: [
+        {
+          type: 'custom',
+          name: 'TInput',
+        },
+      ],
+    },
+  },
+  type: {
+    type: {
+      type: 'string',
+      value: 'not_length',
+    },
+  },
+  requirement: {
+    type: {
+      type: 'custom',
+      name: 'TRequirement',
+    },
+  },
+};